server/requirements.txt
minio
flask
gunicorn
python-dotenv
pymongo